move build_cond to control_flow
